refactor imports and sort utils
