+++ kwave/reconstruction/beamform.py
@@ -1,6 +1,8 @@
 import warnings
+from typing import Tuple, Optional
 
 import numpy as np
+import scipy
 from matplotlib import pyplot as plt
 from scipy.interpolate import interp1d
 from scipy.signal import hilbert
@@ -9,6 +11,7 @@ from uff.position import Position
 
 import kwave.reconstruction.tools as tools
 from kwave.reconstruction.shifted_transform import ShiftedTransform
+from kwave.utils import TicToc, cart2pol, scale_time
 
 
 def beamform(channel_data: ChannelData):
@@ -176,3 +179,90 @@ def focus(kgrid, input_signal, source_mask, focus_position, sound_speed):
 
     warnings.warn("This method is not fully migrated, might be depricated and is untested.", PendingDeprecationWarning)
     return signal_mat
+
+
+def scan_conversion(
+        scan_lines: np.ndarray,
+        steering_angles,
+        image_size: Tuple[float, float],
+        c0,
+        dt,
+        resolution: Optional[Tuple[int, int]]
+) -> np.ndarray:
+    if resolution is None:
+        resolution = (256, 256)  # in pixels
+
+    x_resolution, y_resolution = resolution
+
+    # assign the inputs
+    x, y = image_size
+
+    # start the timer
+    TicToc.tic()
+
+    # update command line status
+    print('Computing ultrasound scan conversion...')
+
+    # extract a_line parameters
+    Nt = scan_lines.shape[1]
+
+    # calculate radius variable based on the sound speed in the medium and the
+    # round trip distance
+    r = c0 * np.arange(1, Nt + 1) * dt / 2  # [m]
+
+    # create regular Cartesian grid to remap to
+    pos_vec_y_new = np.linspace(0, 1, y_resolution) * y - y / 2
+    pos_vec_x_new = np.linspace(0, 1, x_resolution) * x
+    [pos_mat_x_new, pos_mat_y_new] = np.array(np.meshgrid(pos_vec_x_new, pos_vec_y_new, indexing='ij'))
+
+    # convert new points to polar coordinates
+    [th_cart, r_cart] = cart2pol(pos_mat_x_new, pos_mat_y_new)
+
+    # TODO: move this import statement at the top of the file
+    # Not possible now due to cyclic dependencies
+    from kwave.utils.interp import interpolate2d_with_queries
+
+    # below part has some modifications
+    # we flatten the _cart matrices and build queries
+    # then we get values at the query locations
+    # and reshape the values to the desired size
+    # These three steps can be accomplished in one step in Matlab
+    # However, we don't want to add custom logic to the `interpolate2D_with_queries` method.
+
+    # Modifications -start
+    queries = np.array([r_cart.flatten(), th_cart.flatten()]).T
+
+    b_mode = interpolate2d_with_queries(
+        [r, 2 * np.pi * steering_angles / 360],
+        scan_lines.T,
+        queries,
+        method='linear',
+        copy_nans=False
+    )
+    image_size_points = (len(pos_vec_x_new), len(pos_vec_y_new))
+    b_mode = b_mode.reshape(image_size_points)
+    # Modifications -end
+
+    b_mode[np.isnan(b_mode)] = 0
+
+    # update command line status
+    print(f'  completed in {scale_time(TicToc.toc())}')
+
+    return b_mode
+
+
+def envelope_detection(signal):
+    """
+    envelopeDetection applies the Hilbert transform to extract the
+    envelope from an input vector x. If x is a matrix, the envelope along
+    the last axis.
+
+    Args:
+        signal:
+
+    Returns:
+        signal_envelope:
+
+    """
+
+    return np.abs(scipy.signal.hilbert(signal))
