+++ kwave/utils/__init__.py
@@ -1,14 +1,14 @@
-from kwave.utils.checks import *
-from kwave.utils.colormap import *
-from kwave.utils.conversion import *
-from kwave.utils.data import *
-from kwave.utils.dotdictionary import *
-from kwave.utils.filters import *
-from kwave.utils.interp import *
-from kwave.utils.io import *
-from kwave.utils.mapgen import *
-from kwave.utils.math import *
-from kwave.utils.matrix import *
-from kwave.utils.pml import *
-from kwave.utils.signals import *
-from kwave.utils.tictoc import *
+from .checks import *
+from .colormap import *
+from .conversion import *
+from .data import *
+from .dotdictionary import *
+from .filters import *
+from .interp import *
+from .io import *
+from .mapgen import *
+from .math import *
+from .matrix import *
+from .pml import *
+from .signals import *
+from .tictoc import *
