+++ tests/test_filterutils.py
@@ -2,7 +2,9 @@ from math import pi
 
 import numpy as np
 
-from kwave.utils.filters import gaussian, create_cw_signals, fwhm, envelope_detection, sharpness, \
+from kwave.reconstruction.beamform import envelope_detection
+from kwave.utils import create_cw_signals
+from kwave.utils.filters import gaussian, fwhm, sharpness, \
     smooth
 from kwave.utils.signals import get_win
